Move the tests to __tests__ directories
