+++ redux-ship-clarus/src/__tests__/controller.test.js
@@ -1,6 +1,6 @@
 // @flow
-import * as Test from './test';
-import * as Controller from './controller';
+import * as Test from '../test';
+import * as Controller from '../controller';
 
 Test.simulateControl(Controller.control, {
   simpleGifLoad: {"action":{"type":"RandomGif","action":{"type":"Load","tag":"dogs"}},"snapshot":[{"type":"Commit","commit":{"type":"RandomGif","commit":{"type":"LoadStart"}}},{"type":"Effect","effect":{"type":"HttpRequest","url":"http://api.giphy.com/v1/gifs/random?api_key=dc6zaTOxFJmzC&tag=dogs"},"result":"\n\n{\"data\":{\"type\":\"gif\",\"id\":\"5EJHDSPpFhbG0\",\"url\":\"http:\\/\\/giphy.com\\/gifs\\/dogs-exercise-workout-5EJHDSPpFhbG0\",\"image_original_url\":\"http:\\/\\/media1.giphy.com\\/media\\/5EJHDSPpFhbG0\\/giphy.gif\",\"image_url\":\"http:\\/\\/media1.giphy.com\\/media\\/5EJHDSPpFhbG0\\/giphy.gif\",\"image_mp4_url\":\"http:\\/\\/media1.giphy.com\\/media\\/5EJHDSPpFhbG0\\/giphy.mp4\",\"image_frames\":\"60\",\"image_width\":\"296\",\"image_height\":\"157\",\"fixed_height_downsampled_url\":\"http:\\/\\/media1.giphy.com\\/media\\/5EJHDSPpFhbG0\\/200_d.gif\",\"fixed_height_downsampled_width\":\"377\",\"fixed_height_downsampled_height\":\"200\",\"fixed_width_downsampled_url\":\"http:\\/\\/media1.giphy.com\\/media\\/5EJHDSPpFhbG0\\/200w_d.gif\",\"fixed_width_downsampled_width\":\"200\",\"fixed_width_downsampled_height\":\"106\",\"fixed_height_small_url\":\"http:\\/\\/media1.giphy.com\\/media\\/5EJHDSPpFhbG0\\/100.gif\",\"fixed_height_small_still_url\":\"http:\\/\\/media1.giphy.com\\/media\\/5EJHDSPpFhbG0\\/100_s.gif\",\"fixed_height_small_width\":\"189\",\"fixed_height_small_height\":\"100\",\"fixed_width_small_url\":\"http:\\/\\/media1.giphy.com\\/media\\/5EJHDSPpFhbG0\\/100w.gif\",\"fixed_width_small_still_url\":\"http:\\/\\/media1.giphy.com\\/media\\/5EJHDSPpFhbG0\\/100w_s.gif\",\"fixed_width_small_width\":\"100\",\"fixed_width_small_height\":\"53\",\"username\":\"\",\"caption\":\"\"},\"meta\":{\"status\":200,\"msg\":\"OK\",\"response_id\":\"58034bfce70e116b4a19f2d9\"}}"},{"type":"Commit","commit":{"type":"RandomGif","commit":{"type":"LoadSuccess","gifUrl":"http://media1.giphy.com/media/5EJHDSPpFhbG0/giphy.gif"}}}]},
