+++ redux-ship-clarus/src/__tests__/view.test.js
@@ -1,8 +1,8 @@
 // @flow
 import React from 'react';
-import * as Test from './test';
-import * as Model from './model';
-import Index from './view';
+import * as Test from '../test';
+import * as Model from '../model';
+import Index from '../view';
 
 const defaultProps = {
   dispatch: Test.dispatch,
