+++ redux-ship-clarus/src/button/__tests__/view.test.js
@@ -1,8 +1,8 @@
 // @flow
 import React from 'react';
-import * as Test from '../test';
-import * as ButtonModel from './model';
-import Button from './view';
+import * as Test from '../../test';
+import * as ButtonModel from '../model';
+import Button from '../view';
 
 const defaultProps = {
   dispatch: Test.dispatch,
