+++ redux-ship-clarus/src/counter/__tests__/view.test.js
@@ -1,8 +1,8 @@
 // @flow
 import React from 'react';
-import * as Test from '../test';
-import * as CounterModel from './model';
-import Counter from './view';
+import * as Test from '../../test';
+import * as CounterModel from '../model';
+import Counter from '../view';
 
 const defaultProps = {
   dispatch: Test.dispatch,
