+++ redux-ship-clarus/src/random-gif-pair/__tests__/view.test.js
@@ -1,8 +1,8 @@
 // @flow
 import React from 'react';
-import * as Test from '../test';
-import * as RandomGifPairModel from './model';
-import RandomGifPair from './view';
+import * as Test from '../../test';
+import * as RandomGifPairModel from '../model';
+import RandomGifPair from '../view';
 
 const defaultProps = {
   dispatch: Test.dispatch,
