+++ redux-ship-clarus/src/random-gif/__tests__/view.test.js
@@ -1,8 +1,8 @@
 // @flow
 import React from 'react';
-import * as Test from '../test';
-import * as RandomGifModel from './model';
-import RandomGif from './view';
+import * as Test from '../../test';
+import * as RandomGifModel from '../model';
+import RandomGif from '../view';
 
 const defaultProps = {
   dispatch: Test.dispatch,
